Added LinkSwordBeamState for when he uses that throwable
Fixed link walking animation, This should work now hopefully

Started working on Collision detection and Response. I put a ReadMe in the collision folder and tried to put lots of comments so its easy to continue to work on. Most of it commented out at the moment but its a good start.